SONARCLOUD-193 move TypedException to correct package
