+++ server/sonar-ce-task/src/main/java/org/sonar/ce/task/TypedException.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.ce.task.step;
+package org.sonar.ce.task;
 
 /**
  * This interface is implemented by the exceptions
