+++ server/sonar-ce/src/test/java/org/sonar/ce/queue/InternalCeQueueImplTest.java
@@ -36,7 +36,7 @@ import org.sonar.ce.monitoring.CEQueueStatus;
 import org.sonar.ce.monitoring.CEQueueStatusImpl;
 import org.sonar.ce.task.CeTask;
 import org.sonar.ce.task.CeTaskResult;
-import org.sonar.ce.task.step.TypedException;
+import org.sonar.ce.task.TypedException;
 import org.sonar.core.util.UuidFactory;
 import org.sonar.core.util.UuidFactoryImpl;
 import org.sonar.db.DbSession;
